Refactor main.py to separate classes and functions
Separate classes and functions from `main.py` into distinct modules for better organization.

* **Add new files:**
  - Add `models/agent.py` containing the `Agent` class.
  - Add `models/environment.py` containing the `Environment` class.
  - Add `utils/message_utils.py` containing the `Message` class.

* **Modify `main.py`:**
  - Remove the `Agent` class and its methods.
  - Remove the `Environment` class and its methods.
  - Remove the `Message` class and its methods.
  - Import the `Agent`, `Environment`, and `Message` classes from their new locations.

* **Update examples:**
  - Update `examples/senator_example.py` to import `Agent` and `Environment` from their new locations.
  - Update `examples/student_example.py` to import `Agent` and `Environment` from their new locations.
  - Update `examples/teacher_example.py` to import `Agent` and `Environment` from their new locations.
